requirements.txt
fastapi
uvicornsqlmodel
